WIP add new storage column
