+++ crates/fuel-core/Cargo.toml
@@ -43,6 +43,7 @@ hex = { version = "0.4", features = ["serde"] }
 hyper = { workspace = true }
 indicatif = { workspace = true, default-features = true }
 itertools = { workspace = true }
+num_enum = { workspace = true }
 num_cpus = { version = "1.16.0", optional = true }
 rand = { workspace = true }
 rocksdb = { version = "0.21", default-features = false, features = [
